Add documentation for decorators and move synchronized from utils to decorators.
