+++ docs/core/decorators.rst
@@ -1,0 +1,9 @@
+.. _decorators:
+
+
+Decorators
+##########
+
+
+.. automodule:: kinto.core.decorators
+    :members:
