+++ docs/core/index.rst
@@ -15,6 +15,7 @@ like notifications when you write a plugin.
    rationale
    quickstart
    resource
+   decorators
    viewsets
    storage
    cache
