+++ kinto/core/cache/memory.py
@@ -1,6 +1,7 @@
 from kinto import logger
 from kinto.core.cache import CacheBase
-from kinto.core.utils import msec_time, synchronized
+from kinto.core.utils import msec_time
+from kinto.core.decorators import synchronized
 
 
 class Cache(CacheBase):
