+++ kinto/core/decorators.py
@@ -1,3 +1,4 @@
+import threading
 from functools import update_wrapper
 from pyramid.response import Response
 
@@ -17,3 +18,25 @@ class cache_forever(object):
 
         request.response.write(self.saved)
         return request.response
+
+
+def synchronized(method):
+    """Class method decorator to make sure two threads do not execute some code
+    at the same time (c.f Java ``synchronized`` keyword).
+
+    The decorator installs a mutex on the class instance.
+    """
+    def decorated(self, *args, **kwargs):
+        try:
+            lock = getattr(self, '__lock__')
+        except AttributeError:
+            lock = threading.RLock()
+            setattr(self, '__lock__', lock)
+
+        lock.acquire()
+        try:
+            result = method(self, *args, **kwargs)
+        finally:
+            lock.release()
+        return result
+    return decorated
